Cleaned up server abilities + removed user network manager

+++ Assets/Scripts/Game/Server/Abilities/CommandPerformers/ServerAgilityBeginPrimaryFire.cs
@@ -4,7 +4,7 @@ using Game.Common.Gameplay.Ship;
 public class ServerAgilityBeginPrimaryFire: ICommand {
     public bool Receive(ShipManager shipManager, ICommandNetworker networker, CommandPacketData packetData) {
         shipManager.shipLoadout.PrimaryFire.Execute(shipManager, true);
-        networker.SendData(packetData, (int)CommandType.AGILITY_PRIMARY_FIRE_START, shipManager.playerMatchID);
+        networker.SendData(packetData, CommandType.AGILITY_PRIMARY_FIRE_START, shipManager.playerMatchID);
         return true;
     }
 }
