+++ Assets/Scripts/Game/Server/Abilities/CommandPerformers/ServerHeavyEndPrimaryFire.cs
@@ -4,7 +4,7 @@ using Game.Common.Gameplay.Ship;
 public class ServerHeavyEndPrimaryFire: ICommand {
     public bool Receive(ShipManager shipManager, ICommandNetworker networker, CommandPacketData packetData) {
         shipManager.shipLoadout.PrimaryFire.Stop(shipManager, true);
-        networker.SendData(packetData, (int)CommandType.HEAVY_PRIMARY_FIRE_END, shipManager.playerMatchID);
+        networker.SendData(packetData, CommandType.HEAVY_PRIMARY_FIRE_END, shipManager.playerMatchID);
         return true;
     }
 }
